+++ src/main/java/lmzr/photomngr/ui/filter/CopiesComponentFilterUI.java
@@ -27,25 +27,25 @@ public class CopiesComponentFilterUI extends ComponentFilterUI {
         getPane().add(new JLabel("min"));
         a_min = new JFormattedTextField(NumberFormat.getIntegerInstance());
         getPane().add(a_min);
-        a_min.setValue(filter.getMin());
+        a_min.setValue(Long.valueOf(filter.getMin()));
 
         getPane().add(new JLabel("max"));
         a_max = new JFormattedTextField(NumberFormat.getIntegerInstance());
         getPane().add(a_max);
-        a_max.setValue(filter.getMax());
+        a_max.setValue(Long.valueOf(filter.getMax()));
     }
 
     /**
      * @return min value
      */
     public long getMin() {
-        return (Long)a_min.getValue();
+        return ((Long)a_min.getValue()).longValue();
     }
 
     /**
      * @return max value
      */
     public long getMax() {
-        return (Long)a_max.getValue();
+        return ((Long)a_max.getValue()).longValue();
     }
 }
